Add GitHub Actions workflow for code checkout and logging
This commit adds a new file `.github/workflows/test.yaml` which contains a GitHub Actions workflow. The workflow is triggered on push events to the `development` and `main` branches. It includes a job called `checkout-code` that runs on an Ubuntu environment. The job has two steps: one for checking out the code using the `actions/checkout@v3` action, and another for logging the contents of the current folder using the `ls -la` command.